+++ backend/apid/middlewares/allowlist_test.go
@@ -6,18 +6,17 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	"github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/backend/authentication/jwt"
 	"github.com/sensu/sensu-go/backend/store"
 	"github.com/sensu/sensu-go/testing/mockstore"
-	"github.com/sensu/sensu-go/types"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestAllowList(t *testing.T) {
 	// Create a token
-	user := &types.User{Username: "foo"}
-	token, tokenString, _ := jwt.AccessToken(user)
-	claims, _ := jwt.GetClaims(token)
+	claims := v2.FixtureClaims("foo", nil)
+	_, tokenString, _ := jwt.AccessToken(claims)
 
 	store := &mockstore.MockStore{}
 	store.On("GetToken", claims.Subject, claims.Id).Return(claims, nil)
@@ -39,9 +38,8 @@ func TestAllowList(t *testing.T) {
 
 func TestMissingTokenFromAllowList(t *testing.T) {
 	// Create a token
-	user := &types.User{Username: "foo"}
-	token, tokenString, _ := jwt.AccessToken(user)
-	claims, _ := jwt.GetClaims(token)
+	claims := v2.FixtureClaims("bar", nil)
+	_, tokenString, _ := jwt.AccessToken(claims)
 
 	sErr := &store.ErrNotFound{}
 	store := &mockstore.MockStore{}