+++ backend/apid/routers/authentication_test.go
@@ -9,8 +9,12 @@ import (
 	"testing"
 
 	"github.com/gorilla/mux"
+	"github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/backend/apid/middlewares"
 	"github.com/sensu/sensu-go/backend/authentication/jwt"
+	"github.com/sensu/sensu-go/backend/authentication/providers"
+	"github.com/sensu/sensu-go/backend/authentication/providers/basic"
+	realStore "github.com/sensu/sensu-go/backend/store"
 	"github.com/sensu/sensu-go/testing/mockstore"
 	"github.com/sensu/sensu-go/types"
 	"github.com/stretchr/testify/assert"
@@ -19,7 +23,7 @@ import (
 
 func TestLoginNoCredentials(t *testing.T) {
 	store := &mockstore.MockStore{}
-	a := &AuthenticationRouter{store}
+	a := authenticationRouter(store)
 
 	req, _ := http.NewRequest(http.MethodGet, "/auth", nil)
 
@@ -29,7 +33,7 @@ func TestLoginNoCredentials(t *testing.T) {
 
 func TestLoginInvalidCredentials(t *testing.T) {
 	store := &mockstore.MockStore{}
-	a := &AuthenticationRouter{store}
+	a := authenticationRouter(store)
 
 	user := types.FixtureUser("foo")
 	store.
@@ -45,10 +49,10 @@ func TestLoginInvalidCredentials(t *testing.T) {
 
 func TestLoginSuccessful(t *testing.T) {
 	store := &mockstore.MockStore{}
-	a := &AuthenticationRouter{store}
+	a := authenticationRouter(store)
 
 	user := types.FixtureUser("foo")
-	store.On("CreateToken", mock.Anything).Return(nil)
+	store.On("AllowTokens", mock.AnythingOfType("[]*jwt.Token")).Return(nil)
 	store.
 		On("AuthenticateUser", mock.Anything, "foo", "P@ssw0rd!").
 		Return(user, nil)
@@ -72,7 +76,7 @@ func TestLoginSuccessful(t *testing.T) {
 
 func TestTestNoCredentials(t *testing.T) {
 	store := &mockstore.MockStore{}
-	a := &AuthenticationRouter{store}
+	a := authenticationRouter(store)
 
 	req, _ := http.NewRequest(http.MethodGet, "/auth/test", nil)
 
@@ -82,7 +86,7 @@ func TestTestNoCredentials(t *testing.T) {
 
 func TestTestInvalidCredentials(t *testing.T) {
 	store := &mockstore.MockStore{}
-	a := &AuthenticationRouter{store}
+	a := authenticationRouter(store)
 
 	user := types.FixtureUser("foo")
 	store.
@@ -98,7 +102,7 @@ func TestTestInvalidCredentials(t *testing.T) {
 
 func TestTestSuccessful(t *testing.T) {
 	store := &mockstore.MockStore{}
-	a := &AuthenticationRouter{store}
+	a := authenticationRouter(store)
 
 	user := types.FixtureUser("foo")
 	store.
@@ -114,18 +118,15 @@ func TestTestSuccessful(t *testing.T) {
 
 func TestLogoutNotWhitelisted(t *testing.T) {
 	store := &mockstore.MockStore{}
-	a := &AuthenticationRouter{store}
+	a := authenticationRouter(store)
 
 	// Mock calls to the store
-	store.On(
-		"DeleteTokens",
-		mock.AnythingOfType("string"),
-		mock.AnythingOfType("[]string"),
-	).Return(fmt.Errorf("error"))
+	store.On("RevokeTokens", mock.AnythingOfType("[]*v2.Claims")).Return(fmt.Errorf("error"))
 
-	user := &types.User{Username: "foo"}
-	_, tokenString, _ := jwt.AccessToken(user)
-	_, refreshTokenString, _ := jwt.RefreshToken(user)
+	claims := v2.FixtureClaims("foo", nil)
+	_, tokenString, _ := jwt.AccessToken(claims)
+	refreshClaims := &v2.Claims{StandardClaims: v2.StandardClaims(claims.Subject)}
+	_, refreshTokenString, _ := jwt.RefreshToken(refreshClaims)
 	body := &types.Tokens{Refresh: refreshTokenString}
 	payload, _ := json.Marshal(body)
 
@@ -138,18 +139,15 @@ func TestLogoutNotWhitelisted(t *testing.T) {
 
 func TestLogoutSuccess(t *testing.T) {
 	store := &mockstore.MockStore{}
-	a := &AuthenticationRouter{store}
+	a := authenticationRouter(store)
 
 	// Mock calls to the store
-	store.On(
-		"DeleteTokens",
-		mock.AnythingOfType("string"),
-		mock.AnythingOfType("[]string"),
-	).Return(nil)
+	store.On("RevokeTokens", mock.AnythingOfType("[]*v2.Claims")).Return(nil)
 
-	user := &types.User{Username: "foo"}
-	_, tokenString, _ := jwt.AccessToken(user)
-	_, refreshTokenString, _ := jwt.RefreshToken(user)
+	claims := v2.FixtureClaims("foo", nil)
+	_, tokenString, _ := jwt.AccessToken(claims)
+	refreshClaims := &v2.Claims{StandardClaims: v2.StandardClaims(claims.Subject)}
+	_, refreshTokenString, _ := jwt.RefreshToken(refreshClaims)
 	body := &types.Tokens{Refresh: refreshTokenString}
 	payload, _ := json.Marshal(body)
 
@@ -162,7 +160,7 @@ func TestLogoutSuccess(t *testing.T) {
 
 func TestTokenRefreshTokenNotWhitelisted(t *testing.T) {
 	store := &mockstore.MockStore{}
-	a := &AuthenticationRouter{store}
+	a := authenticationRouter(store)
 	user := &types.User{Username: "foo"}
 
 	// Mock calls to the store
@@ -170,14 +168,16 @@ func TestTokenRefreshTokenNotWhitelisted(t *testing.T) {
 		"GetToken",
 		mock.AnythingOfType("string"),
 		mock.AnythingOfType("string"),
-	).Return(&types.Claims{}, fmt.Errorf("error"))
+	).Return(&types.Claims{}, &realStore.ErrNotFound{})
 	store.On("GetUser",
 		mock.AnythingOfType("*context.valueCtx"),
 		mock.AnythingOfType("string"),
 	).Return(user, nil)
 
-	_, tokenString, _ := jwt.AccessToken(user)
-	_, refreshTokenString, _ := jwt.RefreshToken(user)
+	claims := v2.FixtureClaims("foo", nil)
+	_, tokenString, _ := jwt.AccessToken(claims)
+	refreshClaims := &v2.Claims{StandardClaims: v2.StandardClaims(claims.Subject)}
+	_, refreshTokenString, _ := jwt.RefreshToken(refreshClaims)
 	body := &types.Tokens{Refresh: refreshTokenString}
 	payload, _ := json.Marshal(body)
 
@@ -185,33 +185,28 @@ func TestTokenRefreshTokenNotWhitelisted(t *testing.T) {
 	req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", tokenString))
 	res := processRequestWithRefreshToken(a, req)
 
-	assert.Equal(t, http.StatusInternalServerError, res.Code)
+	assert.Equal(t, http.StatusUnauthorized, res.Code)
 }
 
 func TestTokenCannotWhitelistAccessToken(t *testing.T) {
 	store := &mockstore.MockStore{}
-	a := &AuthenticationRouter{store}
+	a := authenticationRouter(store)
 	user := &types.User{Username: "foo"}
 
 	// Mock calls to the store
-	store.On("CreateToken", mock.Anything).Return(fmt.Errorf("error"))
-	store.On(
-		"DeleteTokens",
-		mock.AnythingOfType("string"),
-		mock.AnythingOfType("[]string"),
-	).Return(nil)
-	store.On(
-		"GetToken",
-		mock.AnythingOfType("string"),
-		mock.AnythingOfType("string"),
+	store.On("AllowTokens", mock.AnythingOfType("[]*jwt.Token")).Return(fmt.Errorf("error"))
+	store.On("RevokeTokens", mock.AnythingOfType("[]*v2.Claims")).Return(nil)
+	store.On("GetToken",
+		mock.AnythingOfType("string"), mock.AnythingOfType("string"),
 	).Return(&types.Claims{}, nil)
 	store.On("GetUser",
-		mock.AnythingOfType("*context.valueCtx"),
-		mock.AnythingOfType("string"),
+		mock.AnythingOfType("*context.valueCtx"), mock.AnythingOfType("string"),
 	).Return(user, nil)
 
-	_, tokenString, _ := jwt.AccessToken(user)
-	_, refreshTokenString, _ := jwt.RefreshToken(user)
+	claims := v2.FixtureClaims("foo", nil)
+	_, tokenString, _ := jwt.AccessToken(claims)
+	refreshClaims := &v2.Claims{StandardClaims: v2.StandardClaims(claims.Subject)}
+	_, refreshTokenString, _ := jwt.RefreshToken(refreshClaims)
 	body := &types.Tokens{Refresh: refreshTokenString}
 	payload, _ := json.Marshal(body)
 
@@ -224,28 +219,23 @@ func TestTokenCannotWhitelistAccessToken(t *testing.T) {
 
 func TestTokenSuccess(t *testing.T) {
 	store := &mockstore.MockStore{}
-	a := &AuthenticationRouter{store}
+	a := authenticationRouter(store)
 	user := &types.User{Username: "foo"}
 
 	// Mock calls to the store
-	store.On("CreateToken", mock.Anything).Return(nil)
-	store.On(
-		"DeleteTokens",
-		mock.AnythingOfType("string"),
-		mock.AnythingOfType("[]string"),
-	).Return(nil)
-	store.On(
-		"GetToken",
-		mock.AnythingOfType("string"),
-		mock.AnythingOfType("string"),
+	store.On("AllowTokens", mock.AnythingOfType("[]*jwt.Token")).Return(nil)
+	store.On("RevokeTokens", mock.AnythingOfType("[]*v2.Claims")).Return(nil)
+	store.On("GetToken",
+		mock.AnythingOfType("string"), mock.AnythingOfType("string"),
 	).Return(&types.Claims{}, nil)
 	store.On("GetUser",
-		mock.AnythingOfType("*context.valueCtx"),
-		mock.AnythingOfType("string"),
+		mock.AnythingOfType("*context.valueCtx"), mock.AnythingOfType("string"),
 	).Return(user, nil)
 
-	_, tokenString, _ := jwt.AccessToken(user)
-	_, refreshTokenString, _ := jwt.RefreshToken(user)
+	claims := v2.FixtureClaims("foo", nil)
+	_, tokenString, _ := jwt.AccessToken(claims)
+	refreshClaims := &v2.Claims{StandardClaims: v2.StandardClaims(claims.Subject)}
+	_, refreshTokenString, _ := jwt.RefreshToken(refreshClaims)
 	body := &types.Tokens{Refresh: refreshTokenString}
 	payload, _ := json.Marshal(body)
 
@@ -265,6 +255,13 @@ func TestTokenSuccess(t *testing.T) {
 	assert.NotEqual(t, tokenString, response.Access)
 	assert.NotZero(t, response.ExpiresAt)
 	assert.NotEmpty(t, response.Refresh)
+}
+
+func authenticationRouter(store realStore.Store) *AuthenticationRouter {
+	authenticator := &providers.Authenticator{}
+	provider := &basic.Provider{Store: store}
+	authenticator.AddProvider(provider)
+	return &AuthenticationRouter{store: store, authenticator: authenticator}
 }
 
 func processRequestWithRefreshToken(